tools/game_seed/pyproject.toml
[project]
name = "game-seed"
version = "0.1.0"
description = "Declarative game data seeder for Living Landz"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "O'Brian" }]
dependencies = [
    "psycopg[binary]>=3.1",
    "python-dotenv>=1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.8",
    "pyright>=1.1",
]

[project.scripts]
game-seed = "game_seed.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/game_seed"]

[tool.ruff]
target-version = "py311"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C4",
    "UP",
    "ARG",
    "SIM",
    "RUF",
    "PTH",
    "T20",
]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["game_seed"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
include = ["src"]
exclude = ["**/__pycache__", ".venv"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
